Large changes to bytecode - changing jump ops, no more 8 byte imm jump - added new built in functions - added 8 byte imm move - need to figure out updating jump ops

+++ src/interpreter/interpreter.hpp
@@ -12,7 +12,7 @@ struct Interpreter {
   uint64_t registers [NUM_REGISTERS] {0};
   unsigned char *stack {0};
   unsigned char const * const program;
-  int64_t exitCode {0};
+  int exitCode {0};
   bool z {0}; // zero flag
   bool p {0}; // positive flag
   bool running {true};
@@ -22,5 +22,5 @@ struct Interpreter {
     unsigned char *programData,
     uint64_t stackSize
   );
-  int64_t runProgram();
+  int runProgram();
 };
